requirements.txt
Flask==2.3.2
python-whois==0.7.3
dnspython==2.3.0
